ios test 一个灵狐的模型演示，可旋转模型

+++ Test/test/app_wrap.h
@@ -10,7 +10,7 @@ public:
 	static void uninit();
 
 	//simple input
-	static void move(int dx,int dy);
+	static void move(float dx,float dy);
 };
